Replace get_time function
This function used clock_gettime, which is a POSIX function.
To make it compatible with more platforms, this new function (get_time_millis)
uses clock instead, a standard C function that returns the number of ticks
since the begining of the program. Then it creates a milliseconds stamp with that.