Cast size_t to unsigned long long in error.h
Since the size of size_t is machine dependent,
it may vary between 4 and 8 bytes.
In some machines size_t is a typedef of an
unsigned long and in others, of a unsigned long long.
So to avoid trouble, allways cast the parameter in the printerr macros
to the highest posible size.