Advancements in BTree
I've integrated a few advancements I made weeks ago in another
branch. There's still a lot of work to do in the BTree.
The add funcion seems to "work", more or less, but the remove is
a complete mess.
This is a really difficult structure, but  I'll keep working on
it in the future. I'm sure sooner or later I'll figure it out :)