Ignore standard error when running valgrind-test
A lot of the tests often check if functions treat errors
like they are suposed to.
For example, they call functions with NULL parameters to check if
they print an error message and return an error code.

The problem is, when running valgrind, those messages are annoying,
and get in the way of valgrind's own output.