Remove #include <stdio.h> from graph.h
I don't like having <stdio.h> in the graph header,
it is only used because the print dijkstra and print floyd
functions receive a FILE pointer. So I replaced it with a
void pointer. I know it's more confusing this way, but it
bugs me to have stdio included with graph.h