Remove vector_empty
Now vector_init does not take any max_elements parameter, and
there is no way to directly build a Vector with an initial max size.
This can be achieved using vector_init to make a new vector and
then vector_reserve, to set a max number of elements.
It is more explicit this way, and avoids the posibility of confusing
vector_empty and vector_isempty.