Correct use of assert
Previously, I used to assert that malloc didn't return NULL.
This is bad since assertions may be ignored (e.g. when NDEBUG
is defined). I've changed that, to make sure that if malloc returns
NULL, the error is propagated to the caller.

Also, instead of checking the parameters of the functions to
be correct I assert them.
For example:
vector_append receives a vector and an element. Previously, if one of
them was NULL, I returned NULL_PARAMETER_ERROR. Now I just assert that
they are not NULL.