Change Stack implementation
Previously, the Stack was a Linked List. This was inefficient, since
we had to malloc a node for every element, losing memory locality and
wasting bytes on the "next" pointers.

It's better to use a Vector to store the elements.
When we push to the stack, we append an element to the Vector, and
when we pop, we simply remove the last element of the vector.
These operations have a constant time complexity.