LinkedList: Fix bug in remove_front and remove_back functions
When there was only one element in the list (i.e. head == tail)
the remove_front and remove_back functions didn't set both to NULL.

remove_front only sets the head to NULL, leaving the tail unmodified.
This causes 'double free' and 'use after free' bugs to occur.
Similarly, remove_back only sets the tail to NULL, but not the head.
Same goes for pop_front and pop_back.

We need to handle the case in which head == tail, by setting both
variables to NULL.