[Graph.c] Avoid cleaning garbage values in graph_remove_vertex
Instead of clearing the edges and weights arrays when removing an
element, leave them as garbage, and only clear them when a new
vertex is added. This way, we avoid the effort of clearing the
arrays until they are needed again.

Let's say we remove 1000 vertices from the graph. We don't need to
iterate through the whole edges and weights arrays 1000 times (O n)
Instead, forget about them, and as we start to add new vertices,
populate the arrays with the default values.