Dictionary: Improve dict_redisperse function
In dict_redisperse, we collect all the elements in the dictionary
(the FULL nodes) and add them again after resizing.

We don't need an array of DictionaryNodes to collect those values.
The state is not relevant, we only save the full ones. So we are wasting
1/3 of the memory allocated.

Also, the loop that re-inserts the values in the dictionary is too long
and repetitive (BAD).